getBestMoveScore must return a int

+++ core/src/main/java/com/alonsoruibal/chess/search/SearchEngine.java
@@ -225,7 +225,7 @@ public class SearchEngine implements Runnable {
 		return globalBestMove;
 	}
 
-	public long getBestMoveScore() {
+	public int getBestMoveScore() {
 		return bestMoveScore;
 	}
